# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Swagger Documentation - Integración de Plataforma",
        "description": "L5 Swagger Tablas para la integración de Micro Servicios",
        "contact": {
            "email": "c.rigollet@duocuc.cl"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://awsapi.test/",
            "description": "PCFantasy API Server"
        }
    ],
    "paths": {
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "productos"
                ],
                "summary": "Muestra un producto por id",
                "description": "Retorna un registro por id",
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Nombre producto"
                                        },
                                        "slug": {
                                            "type": "string",
                                            "example": "nombre-producto"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Descripción de producto"
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "1000"
                                        },
                                        "quantity": {
                                            "type": "number",
                                            "example": "45"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2022-05-21T00:43:54.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2022-05-21T00:43:54.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto no existe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "productos"
                ],
                "summary": "Actualiza un producto por id",
                "description": "Retorna el registro actualizado",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "slug": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "quantity": {
                                                "type": "number"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Actualiza Nombre",
                                    "slug": "Actualiza Slug",
                                    "description": "Actualiza Descripción",
                                    "price": "110.000",
                                    "quantity": "45"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Actualizar Nombre"
                                        },
                                        "slug": {
                                            "type": "string",
                                            "example": "actualizar-nombre-de-slug"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Actualizar Descripción"
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "900.000"
                                        },
                                        "quantity": {
                                            "type": "number",
                                            "example": "45"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2022-05-21T00:43:54.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2022-05-21T00:43:54.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "productos"
                ],
                "summary": "Elimina un producto por id",
                "description": "Retorna éxito de la eliminación",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto eliminado con éxito"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}